Add a Temporary Controller Code, only has the basic structure, untested
